third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-border-slice-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-border-slice with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-border-slice">
<meta name="assert" content="mask-border-slice supports only the grammar '<number-percentage>{1,4} fill?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("mask-border-slice", "fill");
test_invalid_value("mask-border-slice", "fill 10%");
test_invalid_value("mask-border-slice", "10px");
test_invalid_value("mask-border-slice", "10% 20 -30% -40 50%");
</script>
</body>
</html>
